Refactored scene widget
